requirements.txt
# Telegram Note Taker 依赖包
python-telegram-bot==20.6
python-dotenv==1.0.0

# AI 总结功能依赖
aiohttp==3.9.0
openai==1.3.0
anthropic==0.7.0